refactor naming and code organization
