+++ lib/wordmove.rb
@@ -16,9 +16,10 @@ require 'photocopier'
 
 require 'wordmove/exceptions'
 require 'wordmove/cli'
+require 'wordmove/doctor'
+require 'wordmove/doctor/movefile'
 require 'wordmove/logger'
 require 'wordmove/movefile'
-require 'wordmove/movefile/doctor'
 require 'wordmove/sql_adapter/default'
 require 'wordmove/sql_adapter/wpcli'
 require "wordmove/version"
